backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
tavily-python==0.5.0
reka-api==3.2.0
